requirements.txt
attrs==20.3.0
coverage==5.4
GitPython==3.1.14
google-auth-oauthlib==0.4.4
iniconfig==1.1.1
joblib==1.0.0
numpy==1.18.5
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
scikit-learn==0.24.1
scipy==1.4.1
tensorflow==2.3.0
toml==0.10.2
yapf==0.30.0
nltk==3.6.2
